requirements.txt
click==6.6
dateutils==0.6.6
enum-compat==0.0.2
eventlet==0.20.0
Flask==0.12
Flask-SocketIO==2.8.2
Flask-WTF==0.13.1
greenlet==0.4.11
itsdangerous==0.24
Jinja2==2.8.1
MarkupSafe==0.23
python-dateutil==2.6.0
python-engineio==1.1.0
python-socketio==1.6.1
pytz==2016.10
six==1.10.0
Werkzeug==0.11.13
WTForms==2.1
